VG-557 Add back some of the lama tests (#7)
This commit adds back some of the lama tests:
- An integration test on the Interpreter. This test uses a real postgres
  in a docker container.
- An end to end on the whole cria worker. This test uses a postgres in
  docker, a keychain in docker and the production explorer. It runs on a
  few test cases on real blockchain: btc, btc_testnet and ltc.

To access the database contents, some of the lama queries and services
have been added back as test queries and services.

Some of the checks needs the keychain id used during the cria run. In
order to access this keychain id, I added a "dump" option to have the
cria worker dump the keychain id to disk. The keychain id is then read
in the tests. This solution is temporary and should be replaced either
by a db table holding the xpub -> keychain id matches or a new
"getKeychain" call on the keychain app.