VG-576 Use keychainId in cria worker (#8)
This commit replaces the (xpub, scheme, lookahead) arguments by a
keychain id. It will now be the orchestrator responsibility to provide
a keychain id from the (xpub, scheme, lookahead) tuple.

This change was made because we had to maintain the xpub -> keychain id
relationship somewhere and we decided it won't be in the cria worker
nor in the keychain. So it will be in the orchestrator.