VG-765 Support double writes to persistence layer (#20)
* VG-765 Support double writes to persistence layer

The goal of this commit is to support writting to both:
- the WD database schema
- the Lama database schema
at the same time in order to gain confidence on our ability to move out
of the WD database when need arises.

To do so, persistence backends can be merged with a "tee" which will
check every access to the secondary backend against the corresponding
access the primary backend. When the backend reads don't match, the
"tee" will log or fail, depending on configuration.

More precisely, writes are not limited to two backends. The persistence
config is one of:
- a wallet daemon persistence config
- a lama persistence config
- a tee config (two leaf configs, on-diff behaviour and concurrency
  setting)
Thus the persistency layer is shaped like a binary tree with any number
of persistence leaves.