VG-872 Add e2e test for incremental sync
This commit adds a test to check that the incremental synchronization
works.
The test synchronizes up to a block (not the latest, some txs are
missing), then synchronizes to the latest and finally checks that the
result is the same as doing only one sync to the latest block.

This commit also moves some e2e tests and fixes some formatting.