VG-869 Support V2 explorer in cria worker (#37)
This commit adds the support for v2 explorer in the worker.
One can choose between `v3` and `v2` in the configuration.
For example:
```
explorer {
  type = explorer-v2
  http {
    uri = "https://explorers.api.vault.ledger.com"
    addresses-size = 20
    txs-batch-size = 500
    timeout = 60 seconds
  }
  fallback {
    type = explorer-v3
    http {
      uri = "https://explorers.api.vault.ledger.com"
      addresses-size = 20
      txs-batch-size = 500
      timeout = 60 seconds
    }
  }
}
```
The `v2` API lacks a route to GET block by hashes, so one needs to
provide a fallback for this method. In the above example, we fallback on
a `v3` instance to get blocks by hash.

Another shortcoming of the `v2` API is that it does not provide
`txinwitness` info in transaction inputs.